# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Productos",
        "description": "Documentación de la API de Productos",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Iniciar sesión y obtener token JWT",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "usuario1@usuario1.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "usuario1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/api/producto": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Lista todos los productos",
                "operationId": "0ae2747cdd1920ea98ae53f68e021678",
                "responses": {
                    "200": {
                        "description": "Lista obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producto"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crea un nuevo producto",
                "operationId": "31a377e56d699faab358339be17a750e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/producto/{id}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtiene un producto por ID",
                "operationId": "af41299eb66ce2d7e6a7c81112895d23",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualiza un producto",
                "operationId": "e295410d9e6c7e46dd5cb53161d676ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Producto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Elimina un producto",
                "operationId": "26f9b1a91c32db076a712a31ed7ddd2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "No se puede eliminar el producto porque está relacionado con otras tablas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/producto/{id}/precios": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtiene los precios de un producto en diferentes divisas",
                "operationId": "7a70bc38bb02a0581c0df6cb2a8c8946",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de precios por divisa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "nombre": {
                                                "type": "string",
                                                "example": "Laptop"
                                            },
                                            "divisa": {
                                                "type": "string",
                                                "example": "USD"
                                            },
                                            "price": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 1200.5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crea un nuevo precio para un producto",
                "operationId": "ce9c1bdaf9298b86997099baf9881857",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "divisa_id",
                                    "price"
                                ],
                                "properties": {
                                    "divisa_id": {
                                        "description": "ID de la divisa",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "price": {
                                        "description": "Precio del producto en la divisa indicada",
                                        "type": "number",
                                        "format": "float",
                                        "example": 1200.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Precio agregado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Precio agregado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Producto": {
                "required": [
                    "name",
                    "price",
                    "divisa_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Laptop"
                    },
                    "description": {
                        "type": "string",
                        "example": "Laptop"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 1200.5
                    },
                    "divisa_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "tax_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 1200.5
                    },
                    "manufacturing_cost": {
                        "type": "number",
                        "format": "float",
                        "example": 1200.5
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Productos",
            "description": "Productos"
        }
    ]
}